+++ js/app.js
@@ -1,0 +1,3 @@
+import initValue from "./util/initValue.js";
+
+initValue();
