+++ main.html
@@ -188,5 +188,5 @@
       </section>
     </div>
   </body>
-  <script type="module" src="/js/script.js"></script>
+  <script type="module" src="/js/app.js"></script>
 </html>
